requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.2.2
matplotlib==3.7.2
seaborn==0.12.2
xgboost==3.0.0
streamlit==1.26.0
shap==0.47.0
joblib==1.3.2
imbalanced-learn==0.10.1
pillow>=7.1.0,<10.0.0
openai==0.28.0
python-dotenv==1.0.0
tqdm==4.65.0
